// types.ts
export enum AppStage {
  LOGIN = 'LOGIN',
  CELEBRATION = 'CELEBRATION',
  GIFT_SELECTION = 'GIFT_SELECTION',
  LANDMINE = 'LANDMINE',
  PROPOSAL = 'PROPOSAL',
  SIGNATURE = 'SIGNATURE',
  OFFICIAL_COUPLE = 'OFFICIAL_COUPLE',
  MAZE = 'MAZE',
  FINALE = 'FINALE'
}

export interface Position {
  x: number;
  y: number;
}